fix: Resolve GUI compilation errors
- Fix float64 to float32 conversion in animated_score.go
- Remove unused ramTotalBar variable in monitor_panel.go
- Remove unused storage import in priority_panel.go

+++ gui/views/animated_score.go
@@ -152,7 +152,7 @@ func (r *scoreRingRenderer) Refresh() {
 	// The ring appears to fill as the score increases
 	percentage := r.score.displayScore / 100.0
 	if percentage > 0 {
-		r.ring.StrokeWidth = 10 + (percentage * 5) // Thicker stroke as score increases
+		r.ring.StrokeWidth = float32(10 + (percentage * 5)) // Thicker stroke as score increases
 	}
 
 	canvas.Refresh(r.ring)
